Changes are now applied to the class located at qiskit_nature/second_q/problems
